requirements.txt
fastapi~=0.111.0
uvicorn
sqlalchemy~=2.0.30
psycopg2-binary
python-dotenv~=1.0.1